// src/chatgpt/types.ts
export interface SendResponseOptions<T = any> {
    type: 'Success' | 'Fail'
    message?: string
    data?: T
}